requirements.txt
dnspython>=2.6.0
httpx>=0.27.0
psutil>=5.9.0
numpy>=1.26.0
scipy>=1.12.0
matplotlib>=3.8.0
